% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGKnock.R
\docType{data}
\name{Example.GeneScan3D.UKB.GLMM}
\alias{Example.GeneScan3D.UKB.GLMM}
\alias{GeneScan3D.UKB.GLMM.example}
\title{Data example for GeneScan3D UKBB gene-based testing based on GLMM.}
\format{
An object of class \code{list} of length 8.
}
\usage{
data("GeneScan3D.UKB.GLMM.example")
}
\description{
This example dataset contains a fitted null GLMM results, variance ratio and sparse Sigma matrix, which are obtained by SAIGE/SAIGE-Gene package. Besides, it also contains the original and knockoff genotype for gene and buffer region and enhancer, as well as the positions of gene buffer region.
}
\examples{
data("GeneScan3D.UKB.GLMM.example")

result.null.model.GLMM=GeneScan3D.UKB.GLMM.example$result.null.model.GLMM
ratio=GeneScan3D.UKB.GLMM.example$ratio
sparseSigma=GeneScan3D.UKB.GLMM.example$sparseSigma
G_gene_buffer=GeneScan3D.UKB.GLMM.example$G_gene_buffer
pos_gene_buffer=GeneScan3D.UKB.GLMM.example$pos_gene_buffer
G_gene_buffer_knockoff1=GeneScan3D.UKB.GLMM.example$G_gene_buffer_knockoff1
G_enhancer=GeneScan3D.UKB.GLMM.example$G_enhancer
G_enhancer_knockoff1=GeneScan3D.UKB.GLMM.example$G_enhancer_knockoff1
Gsub.id=result.null.model.GLMM$sampleID
}
\keyword{data}
