% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knockoff.generation.R
\docType{data}
\name{Example.Knock.generation.enhancer}
\alias{Example.Knock.generation.enhancer}
\alias{Knock.generation.enhancer.example}
\title{Data example for knockoff generation of enhancer}
\format{
An object of class \code{list} of length 2.
}
\usage{
data("Knock.generation.enhancer.example")
}
\description{
This example dataset contains a genotype matrix of enhancer surrounding region and the positions for enhancer. The example dataset is extracted from the 'plinkforGRM_1000samples_10kMarkers' plink data in the extdata folder of SAIGE package, which contains 1000 samples.
}
\examples{
data('Knock.generation.enhancer.example')
G_enhancer_surround=Knock.generation.enhancer.example$G_enhancer_surround
pos_enhancer=Knock.generation.enhancer.example$pos_enhancer
}
\keyword{data}
