% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knockoff.generation.R
\docType{data}
\name{Example.Knock.generation.gene.buffer}
\alias{Example.Knock.generation.gene.buffer}
\alias{Knock.generation.gene.buffer.example}
\title{Data example for knockoff generation of gene buffer region}
\format{
An object of class \code{list} of length 2.
}
\usage{
data("Knock.generation.gene.buffer.example")
}
\description{
This example dataset contains a genotype matrix of gene buffer surrounding region and the positions for gene buffer region. The example dataset is extracted from the 'plinkforGRM_1000samples_10kMarkers' plink data in the extdata folder of SAIGE package, which contains 1000 samples.
}
\examples{
data('Knock.generation.gene.buffer.example')
G_gene_buffer_surround=Knock.generation.gene.buffer.example$G_gene_buffer_surround
pos_gene_buffer=Knock.generation.gene.buffer.example$pos_gene_buffer
}
\keyword{data}
