% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knockoff.generation.R
\name{Knockoffgeneration.enhancer}
\alias{Knockoffgeneration.enhancer}
\title{Knockoff Generation of enhancer}
\usage{
Knockoffgeneration.enhancer(
  G_enhancer_surround = G_enhancer_surround,
  enhancer_start = enhancer_start,
  enhancer_end = enhancer_start,
  M = 5,
  surround.region = 50000,
  LD.filter = 0.75
)
}
\arguments{
\item{G_enhancer_surround}{The genotype matrix of the surrounding region for enhancer.}

\item{enhancer_start}{The start position of enhancer.}

\item{enhancer_end}{The end position of enhancer.}

\item{M}{Numer of multiple knockoffs.}

\item{surround.region}{Surrounding region for enhancer, default is +-50kb.}

\item{LD.filter}{The correlation threshold for hierarchical clustering. Default LD filter at correlation 0.75}
}
\value{
{A list file contains original genotype and M knockoff genotypes for enhancer.}
}
\description{
This function generates multiple knockoff genotypes for enhancer. The knockoff generations are optimized using shrinkage leveraging algorithm.
}
\examples{
library(Matrix)
data('Knock.generation.enhancer.example')
G_enhancer_surround=Matrix(Knock.generation.enhancer.example$G_enhancer_surround)
pos_enhancer=Knock.generation.enhancer.example$pos_enhancer
enhancer_start=min(pos_enhancer)
enhancer_end=max(pos_enhancer)

G_enhancer_knockoff=Knockoffgeneration.enhancer(G_enhancer_surround=G_enhancer_surround,
                                             enhancer_start=enhancer_start,
                                             enhancer_end=enhancer_end,
                                             M=5,surround.region=50000,LD.filter=0.75)
#M=5 knockoffs
G_enhancer_knockoff1=G_enhancer_knockoff[1,,]
G_enhancer_knockoff2=G_enhancer_knockoff[2,,]
G_enhancer_knockoff3=G_enhancer_knockoff[3,,]
G_enhancer_knockoff4=G_enhancer_knockoff[4,,]
G_enhancer_knockoff5=G_enhancer_knockoff[5,,]

}
