% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knockoff.generation.R
\name{Knockoffgeneration.gene.buffer}
\alias{Knockoffgeneration.gene.buffer}
\title{Knockoff Generation of Gene buffer region}
\usage{
Knockoffgeneration.gene.buffer(
  G_gene_buffer_surround = G_gene_buffer_surround,
  gene_buffer_start = gene_buffer_start,
  gene_buffer_end = gene_buffer_end,
  M = 5,
  surround.region = 1e+05,
  LD.filter = 0.75
)
}
\arguments{
\item{G_gene_buffer_surround}{The genotype matrix of the surrounding region for gene buffer region.}

\item{gene_buffer_start}{The start position of gene buffer region.}

\item{gene_buffer_end}{The end position of gene buffer region.}

\item{M}{Numer of multiple knockoffs.}

\item{surround.region}{Surrounding region for gene buffer region, default is +-100kb.}

\item{LD.filter}{The correlation threshold for hierarchical clustering. Default LD filter at correlation 0.75}
}
\value{
\item{G_gene_buffer_knockoff}{A list file contains M knockoff genotypes for gene buffer region.}
}
\description{
This function generates multiple knockoff genotypes for gene buffer region. The knockoff generations are optimized using shrinkage leveraging algorithm.
}
\examples{
library(Matrix)
data('Knock.generation.gene.buffer.example')
G_gene_buffer_surround=Matrix(Knock.generation.gene.buffer.example$G_gene_buffer_surround)
pos_gene_buffer=Knock.generation.gene.buffer.example$pos_gene_buffer
gene_buffer_start=min(pos_gene_buffer)
gene_buffer_end=max(pos_gene_buffer)

G_gene_buffer_knockoff=Knockoffgeneration.gene.buffer(G_gene_buffer_surround=G_gene_buffer_surround,
                                               gene_buffer_start=gene_buffer_start,
                                               gene_buffer_end=gene_buffer_end,
                                               M=5,surround.region=100000,LD.filter=0.75)
#M=5 knockoffs
G_gene_buffer_knockoff1=G_gene_buffer_knockoff[1,,]
G_gene_buffer_knockoff2=G_gene_buffer_knockoff[2,,]
G_gene_buffer_knockoff3=G_gene_buffer_knockoff[3,,]
G_gene_buffer_knockoff4=G_gene_buffer_knockoff[4,,]
G_gene_buffer_knockoff5=G_gene_buffer_knockoff[5,,]

}
